added_onboarding component
